What's going on party people! Did some work with the database making sure to test as many "new user" possibilities as I could think of. (example: user account already exists, passwords don't match, security questions don't match, not all fields are filled in).
Left a little utility line in Register Activity. (Line 35 I think when it
comes to testing data for the UserStorage Database). We can delete the
contents of the database when necessary by commenting that line back in.

Starting working on the stubs for the LoginActivity to mimic the database
query to determine validity of user credentials.

Didn't get a chance to start on the custom theme for our app, but will try
to start on that soon.

Hope y'all are having an excellent saturday.